add bootstrap and table responsive class
add bootstrap from official bootstrap cdn. Could also be implemented as local file on the webserver (just point to a bootstrap.min.css and js).
Just made the table to a responsive bootstrap table so you can scroll through the table on small devices. Just improves the mobile usability and does not effect the desktop usability.